Chore extract Swagger UI from utoipa
* Extract Swagger UI from utoipa crate. -> Lots of benefits like smaller crates & better separation of concern
* Update build pipeline to use matix -> Tests now all crates
* Updated docs and README
* Updated Swagger UI to new version 4.6.1
* Updated docs and versions
