+++ .github/workflows/build.yaml
@@ -2,18 +2,33 @@ name: Utoipa build
 
 on:
   push:
+    paths:
+      - "**.rs"
+      - "**Cargo.toml"
   pull_request:
     branches: [ master ]
-
+    paths:
+      - "**.rs"
+      - "**Cargo.toml"
 env:
   CARGO_TERM_COLOR: always
 
 jobs:
   test:
+    strategy:
+      matrix:
+        testset: 
+          - default
+          - actix
+          - gen
+          - swagger
+      fail-fast: true
     runs-on: ubuntu-latest
 
     steps:
     - uses: actions/checkout@v2
+      with:
+        fetch-depth: 2
 
     - uses: actions/cache@v2
       with:
@@ -25,27 +40,35 @@ jobs:
           target/
         key: ${{ runner.os }}-cargo-${{ hashFiles('**/Cargo.lock') }}
     
+    - name: Resolve changed paths
+      id: changes
+      run: |
+        root_changed=false
+        gen_changed=false
+        swagger_changed=false
+        while read -r change; do
+          if [[ "$change" == "utoipa-gen" ]]; then
+            gen_changed=true
+          elif [[ "$change" == "utoipa-swagger-ui" ]]; then
+            swagger_changed=true
+          else
+            root_changed=true
+          fi
+        done < <(git diff --name-only ${{ github.sha }}~ ${{ github.sha }} | grep .rs | awk -F \/ '{print $1}')
+        echo "::set-output name=root_changed::$root_changed"
+        echo "::set-output name=gen_changed::$gen_changed"
+        echo "::set-output name=swagger_changed::$swagger_changed"
+
     - name: Run tests
       run: |
-        cargo test
-        cargo test --test path_response_derive_test_no_serde_json --no-default-features
-        cargo test --test component_derive_no_serde_json --no-default-features
-
-  test-with-actix:
-    runs-on: ubuntu-latest
-
-    steps:
-    - uses: actions/checkout@v2
-
-    - uses: actions/cache@v2
-      with:
-        path: |
-          ~/.cargo/bin/
-          ~/.cargo/registry/index/
-          ~/.cargo/registry/cache/
-          ~/.cargo/git/db/
-          target/
-        key: ${{ runner.os }}-cargo-${{ hashFiles('**/Cargo.lock') }}-actix
-    
-    - name: Run tests
-      run: cargo test --features actix-web,actix_extras
+        if [[ "${{ matrix.testset }}" == "default" ]] && [[ ${{ steps.changes.outputs.root_changed }} == true ]]; then
+          cargo test
+          cargo test --test path_response_derive_test_no_serde_json --no-default-features
+          cargo test --test component_derive_no_serde_json --no-default-features
+        elif [[ "${{ matrix.testset }}" == "actix" ]] && [[ ${{ steps.changes.outputs.root_changed }} == true ]]; then
+          cargo test --features actix_extras
+        elif [[ "${{ matrix.testset }}" == "gen" ]] && [[ ${{ steps.changes.outputs.gen_changed }} == true ]]; then
+          cargo test -p utoipa-gen --features actix_extras
+        elif [[ "${{ matrix.testset }}" == "swagger" ]] && [[ ${{ steps.changes.outputs.swagger_changed }} == true ]]; then
+          cargo test -p utoipa-swagger-ui --features actix-web
+        fi
