+++ Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "utoipa"
 description = "Compile time rest API documentation for Rust"
-version = "0.1.0-beta.6"
+version = "0.1.0-beta6"
 edition = "2021"
 license = "MIT OR Apache-2.0"
 readme = "README.md"
@@ -17,7 +17,6 @@ exclude = [
 
 [features]
 default = ["json"]
-swagger_ui = ["rust-embed", "mime_guess"]
 debug = ["utoipa-gen/debug"]
 actix_extras = ["utoipa-gen/actix_extras"]
 json = ["serde_json", "utoipa-gen/json"]
@@ -25,21 +24,20 @@ json = ["serde_json", "utoipa-gen/json"]
 [dependencies]
 serde = { version = "1.0", features = ["derive"] }
 serde_json = { version = "1.0", optional = true }
-rust-embed = { version = "6.3", optional = true, features = ["interpolate-folder-path"] }
-actix-web =  { version = "3.3", optional = true }
-mime_guess = { version = "2.0", optional = true }
-utoipa-gen = { version = "0.1.0-beta.6", path = "./utoipa-gen" }
+utoipa-gen = { version = "0.1.0-beta6", path = "./utoipa-gen" }
 
 [dev-dependencies]
+actix-web = { version = "3.3" }
 paste = "1.0.6"
 
 [workspace]
 members = [
-  "utoipa-gen"
+  "utoipa-gen",
+  "utoipa-swagger-ui"
 ]
 
 [package.metadata.docs.rs]
-features = ["swagger_ui", "json", "actix-web", "actix_extras"]
+features = ["json", "actix_extras"]
 
 [package.metadata.publish]
-order = ["utoipa-gen", "utoipa"]
+order = ["utoipa-gen", "utoipa", "utoipa-swagger-ui"]
