+++ utoipa-gen/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "utoipa-gen"
 description = "Code generation implementation for utoipa"
-version = "0.1.0-beta.6"
+version = "0.1.0-beta6"
 edition = "2021"
 license = "MIT OR Apache-2.0"
 readme = "README.md"
