+++ utoipa-swagger-ui/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "utoipa-swagger-ui"
+description = "Swagger UI for utoipa"
+version = "0.1.0-beta1"
+edition = "2021"
+license = "MIT OR Apache-2.0"
+readme = "README.md"
+keywords = ["swagger-ui", "openapi", "documentation"]
+repository = "https://github.com/juhaku/utoipa"
+categories = ["web-programming"]
+
+[features]
+debug = []
+
+[dependencies]
+rust-embed = { version = "6.3", features = ["interpolate-folder-path"] }
+mime_guess = { version = "2.0" }
+actix-web =  { version = "3.3", optional = true }
+utoipa = { version = "0.1.0-beta6", path = "..", default-features = false, features = [] }
+
+[package.metadata.docs.rs]
+features = ["actix-web"]
